evez-ecosystem/evezart-repos/evez666-arg-canon/evez_ouroboros/requirements.txt
httpx>=0.27
stripe>=9.0
openai>=1.30
groq>=0.9
mem0ai>=0.1
ably>=2.0
cloudconvert>=3.0
python-dotenv>=1.0
